requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
et-xmlfile==1.0.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
ipython==7.20.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
moviepy==1.0.3
numpy==1.20.1
opencv-python==4.5.1.48
openpyxl==3.0.6
pandas==1.2.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
proglog==0.1.9
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.56.2
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
